refactor task result statistics
